refactor: change project layout
