+++ lib/gulp/content/content-handler.js
