+++ lib/gulp/content/generator.js
@@ -6,7 +6,7 @@ const pug = require('pug')
 const Vinyl = require('vinyl')
 
 const {MiddlewareFactory} = require('./middleware')
-const {async} = require('../utils')
+const {async} = require('../../utils')
 
 const defaultTemplate = pug.compile(`doctype html
 html(lang='en' dir='ltr')
@@ -22,7 +22,7 @@ function _middlewares(state) {
 	return ({files, partials, templates}) => {
 		return async.waterfall([
 			always({files, partials, templates}),
-			...middlewares
+			...middlewares,
 		])
 	}
 }
@@ -41,13 +41,12 @@ function _templateResolver(state, templates) {
 		state.log.info(`No template found for ${slug}`)
 		return defaultTemplate
 	}
-
 }
 
 function _partialsResolver(state, partials) {
 	return slug => pick(
 		Object.keys(partials).filter(startsWith(slug)),
-		partials
+		partials,
 	)
 }
 
@@ -55,7 +54,7 @@ function _makeOutputPath(file) {
 	const {name, dir} = parse(file.path)
 	return join(
 		dir,
-		...(name === 'index' ? [`${name}.html`] : [name, 'index.html'])
+		...(name === 'index' ? [`${name}.html`] : [name, 'index.html']),
 	)
 }
 
@@ -90,7 +89,7 @@ function _htmlize(state) {
 	return ({files, partials, templates}) => {
 		return async.map(
 			_fileRenderer(state, partials, templates),
-			Object.entries(files)
+			Object.entries(files),
 		)
 	}
 }
@@ -127,7 +126,7 @@ function _index(state) {
 	return identity
 }
 
-module.exports = function _Blender(state) {
+module.exports = function _Generator(state) {
 	const {files, partials, templates} = state
 	return stream => async.waterfall([
 		always({files, partials, templates}),
