+++ lib/gulp/content/index.js
@@ -1,5 +1,7 @@
-const Blender = require('../content-generator')
-const {Logger} = require('../utils')
+const {Logger, stream: {Transformer}} = require('../../utils')
+
+const Generator = require('./generator')
+const Loader = require('./loader')
 
 const del = require('del')
 const gulp = require('gulp')
@@ -21,13 +23,19 @@ module.exports = function content({
 	const state = {
 		markdown,
 		middlewares,
-		middlewaresDirectories,
+		middlewaresDirectories: [
+			'',
+			...middlewaresDirectories,
+		],
 		contentsBaseDir: defaultTo(join(baseDir, 'contents'), contentsBaseDir),
 		layoutBaseDir: defaultTo(join(baseDir, 'layout'), layoutBaseDir),
+		outputDirectory: join(outputDirectory, prefix),
 		log: Logger({prefix: 'Blender'}),
 		index,
+		files: {},
+		partials: {},
+		templates: {},
 	}
-	const contentOutputDirectory = join(outputDirectory, prefix)
 	const sources = [
 		join(state.contentsBaseDir, '**', '*.md'),
 		join(state.layoutBaseDir, '**', '*.pug'),
@@ -35,12 +43,15 @@ module.exports = function content({
 	return {
 		build(cb) {
 			return gulp.src(sources)
-				.pipe(Blender(state))
+				.pipe(Transformer({
+					handle: Loader(state),
+					finish: Generator(state),
+				}))
 				.on('error', cb)
-				.pipe(gulp.dest(contentOutputDirectory))
+				.pipe(gulp.dest(state.outputDirectory))
 		},
 		clean(cb) {
-			return del(join(contentOutputDirectory, '**/*.html'))
+			return del(join(state.outputDirectory, '**/*.html'))
 		},
 		sources,
 	}
