+++ lib/gulp/content/layout-handler.js
@@ -4,7 +4,7 @@ const {relative} = require('path')
 
 const {pathToSlug} = require('./slug')
 
-const Logger = require('../utils/logger')
+const Logger = require('../../utils/logger')
 
 function _LayoutHandler(state) {
 	return ({file, ext, name, encoding}) => {
