+++ lib/gulp/content/loader.js
