+++ lib/gulp/content/middleware.js
@@ -1,6 +1,6 @@
 const {cond, curry, either, identity, is, isNil, pipe, T, where} = require('ramda')
 
-const {tap: asyncTap} = require('../utils/async')
+const {tap: asyncTap} = require('../../utils/async')
 
 const isFunction = is(Function)
 const isObject = is(Object)
