+++ lib/gulp/content/slug.js
