+++ tests/specs/content-generator/middleware/middleware-builder.js
@@ -5,7 +5,7 @@ const {spy, stub} = require('sinon')
 const State = require('../helpers/state')
 
 const {noop} = require('../../../../lib/utils/functional')
-const {_MiddlewareBuilder} = require('../../../../lib/content-generator/middleware')
+const {_MiddlewareBuilder} = require('../../../../lib/gulp/content/middleware')
 
 test('_MiddlewareBuilder should be a function', t => {
 	t.true(is(Function, _MiddlewareBuilder))
